backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
websockets==12.0
aiofiles==23.2.1
httpx==0.27.0
faster-whisper==1.0.1
scenedetect[opencv]==0.6.3
librosa==0.10.2
opencv-python==4.9.0.80
deepface==0.0.91
basicsr==1.4.2
realesrgan==0.3.0
Pillow==10.3.0
numpy==1.26.4
redis==5.0.4
python-dotenv==1.0.1
psutil==5.9.8
